pyproject.toml
[project]
name = "cto-sidekick"
version = "0.1.0"
description = "Autonomous development orchestration service"
requires-python = ">=3.10"
dependencies = [
    "gspread>=6.0.0",
    "oauth2client>=4.1.3",
    "pyyaml>=6.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
